# Reduce memory usage when parsing `BigInt` and `BigDecimal` values with JVMs